breakout scoring in hands.go
