+++ cards/hands.go
@@ -1,7 +1,6 @@
 package cards
 
 import (
-	"fmt"
 	"log/slog"
 )
 
@@ -26,6 +25,11 @@ func (h *Hand) IsCompleted() bool {
 	return h.completed
 }
 
+// True if a 2-card hand adds up to 21
+func (h Hand) IsBlackjack() bool {
+	return len(h.Cards) == 2 && h.GetTotalValue() == 21
+}
+
 // Checks that hand length is 2 & for a pair
 func (h Hand) IsSplittable() bool {
 	if len(h.Cards) != 2 ||
@@ -47,50 +51,6 @@ func SplitHand(hand Hand) []Hand {
 		Cards: []Card{hand.Cards[1]},
 	}}
 	return result
-}
-
-// Checks totalValue not > 21
-func (h *Hand) CheckBust() bool {
-	if h.GetTotalValue() > 21 {
-		fmt.Sprintln(fmt.Sprintf("Oof, %d means bust!", h.GetTotalValue()))
-		h.bust = true
-		return true
-	}
-	return false
-}
-
-func (h *Hand) GetTotalValue() int {
-	if h.totalValue != nil {
-		return *h.totalValue
-	}
-
-	totalValue := h.CalculateTotalValue()
-
-	return totalValue
-}
-
-// calculate total & cache result
-func (h *Hand) CalculateTotalValue() int {
-	aceCounter := 0
-	totalValue := 0
-
-	for _, card := range h.Cards {
-		if card.Name == "Ace" {
-			aceCounter++
-		}
-		totalValue += card.Value
-	}
-
-	for totalValue > 21 && aceCounter > 0 {
-		aceCounter--
-		totalValue -= 10
-	}
-
-	// cache & return result
-	h.totalValue = &totalValue
-	h.CheckBust()
-
-	return totalValue
 }
 
 func getBet()          {}
